# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Absensi Attendance API
  version: 1.0.0
  description: API untuk absensi mobile, HR admin, izin/cuti, laporan, dan payroll.
servers:
  - url: http://localhost:8080/api
paths:
  /auth/register:
    post:
      summary: Register karyawan baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User dibuat dan link verifikasi dikirim
  /auth/login:
    post:
      summary: Login dan terima access token + refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token valid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, nip, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        nip:
          type: string
        departmentId:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...